requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
gherkin-official==29.0.0
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
parse==1.20.2
parse_type==0.6.4
pluggy==1.6.0
pytest==8.3.5
pytest-bdd==8.1.0
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
behave==1.2.5
python-dotenv==1.0.1